feat(build): add cross-platform build support for Darwin, Linux, and Windows
