+++ internal/command/root.go
@@ -20,7 +20,6 @@ import (
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"golang.org/x/sync/errgroup"
-	"golang.org/x/sys/unix"
 )
 
 var resizeDisk bool
@@ -84,7 +83,10 @@ func run(cmd *cobra.Command, args []string) error {
 	if runtime.GOOS == "darwin" {
 		version, ok := tart.Version()
 		if !ok {
-			return unix.Kill(os.Getppid(), syscall.SIGTERM)
+			if p, err := os.FindProcess(os.Getppid()); err == nil {
+				_ = p.Signal(syscall.SIGTERM)
+			}
+			return errors.New("failed to identify Tart version on macOS guest")
 		}
 
 		zap.S().Infof("running on Tart %s, proceeding...", version.String())
