+++ internal/execuser/execuser_windows.go
@@ -1,0 +1,12 @@
+//go:build windows
+
+package execuser
+
+import "fmt"
+
+func Resolve(spec string) (any, error) {
+	if spec == "" {
+		return nil, nil
+	}
+	return nil, fmt.Errorf("user switching is not supported on Windows")
+}
