+++ internal/vsock/listener_unix.go
@@ -1,3 +1,5 @@
+//go:build darwin || linux
+
 package vsock
 
 import (
